modify clients styles, lift states up
