+++ src/pages/Clients/ClientsList.jsx
@@ -1,8 +1,5 @@
-import React, { useEffect, useState } from 'react';
-import { connect } from 'react-redux';
-import { FormControlLabel, Switch } from '@material-ui/core';
-import { getClientsAct } from '../../state/clients/clientActions';
-import isEqual from 'lodash.isequal';
+import React from 'react';
+
 import { TableComponent } from '../../components/Table/index';
 
 const headCells = [
@@ -13,58 +10,18 @@ const headCells = [
 ];
 
 const ClientsList = props => {
-    const { clients, isLoggedIn, getClientsAct } = props;
-    const [clientsList, setClientsList] = useState([]);
-    const [dense, setDense] = useState(false);
-
-    useEffect(() => {
-        getClientsAct();
-    }, [getClientsAct]);
-
-    useEffect(() => {
-        if (!isLoggedIn) {
-            const localClients = JSON.parse(
-                window.localStorage.getItem('clients'),
-            );
-            if (!isEqual(localClients, clientsList)) {
-                setClientsList(localClients);
-            }
-        } else {
-            if (!isEqual(clients, clientsList)) {
-                setClientsList(clients);
-            }
-        }
-    }, [isLoggedIn, clients, clientsList, getClientsAct]);
-
-    const handleChangeDense = event => {
-        setDense(event.target.checked);
-    };
+    const { clientsList, selected, setSelected, dense } = props;
 
     return (
-        <>
-            <FormControlLabel
-                control={
-                    <Switch checked={dense} onChange={handleChangeDense} />
-                }
-                label="Dense padding"
-            />
-            <TableComponent
-                data={clientsList}
-                dataType="Client"
-                dense={dense}
-                headCells={headCells}
-            />
-        </>
+        <TableComponent
+            data={clientsList}
+            dataType="Client"
+            dense={dense}
+            headCells={headCells}
+            selected={selected}
+            setSelected={setSelected}
+        />
     );
 };
 
-const mapStateToProps = state => {
-    return {
-        clients: state.clients.clients,
-        isLoggedIn: state.user.isLoggedIn,
-    };
-};
-
-export default connect(mapStateToProps, {
-    getClientsAct,
-})(ClientsList);
+export default ClientsList;
